app/requirements.txt
# Python 3.12.6
python-dotenv
boto3
pytz
tqdm
configparser
tenacity
psutil